requirements.txt
streamlit
fastapi
uvicorn
gitpython
black
pylint
openai
crewai
python-dotenv
langchain-community